% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{kernel_from_dist}
\alias{kernel_from_dist}
\title{Build a phylogenetic kernel from a distance matrix}
\usage{
kernel_from_dist(dist, alpha, kernel = c("gaussian", "laplacian"))
}
\arguments{
\item{dist}{A taxa x taxa distance matrix.}

\item{alpha}{Positive bandwidth parameter.}

\item{kernel}{Either "gaussian" or "laplacian".}
}
\value{
A taxa x taxa kernel matrix.
}
\description{
Build a phylogenetic kernel from a distance matrix
}
\keyword{internal}
