% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-fit.R
\name{phymapnet_precision}
\alias{phymapnet_precision}
\title{Internal PhyMapNet precision MAP estimator}
\usage{
phymapnet_precision(Y, nue, epsilon1, epsilon2, C)
}
\description{
Internal PhyMapNet precision MAP estimator
}
\keyword{internal}
